+++ dev/benchmarks/macrobenchmarks/lib/src/web/bench_mouse_region_mixed_grid_hover.dart
@@ -58,7 +58,7 @@ class BenchMouseRegionMixedGridHover extends WidgetRecorder {
   late _Tester _tester;
 
   void handleDataPoint(Duration duration) {
-    profile!.addDataPoint('hitTestDuration', duration, reported: true);
+    profile?.addDataPoint('hitTestDuration', duration, reported: true);
   }
 
   // Use a non-trivial border to force Web to switch painter
@@ -157,6 +157,8 @@ class _Tester {
   static const Duration hoverDuration = Duration(milliseconds: 20);
 
   bool _stopped = false;
+  final Completer<void> _finished = Completer<void>();
+  Future<void> get finished => _finished.future;
 
   TestGesture get gesture {
     return _gesture ??= TestGesture(
@@ -188,6 +190,7 @@ class _Tester {
       await _hoverTo(const Offset(370, 390), hoverDuration);
       await _hoverTo(const Offset(390, 30), hoverDuration);
     }
+    _finished.complete();
   }
 
   void stop() {
