+++ engine/src/flutter/lib/web_ui/lib/src/engine.dart
@@ -74,7 +74,6 @@ export 'engine/layer/layer_scene_builder.dart';
 export 'engine/layer/layer_tree.dart';
 export 'engine/layer/layer_visitor.dart';
 export 'engine/layer/n_way_canvas.dart';
-export 'engine/layers.dart';
 export 'engine/lazy_path.dart';
 export 'engine/mouse/context_menu.dart';
 export 'engine/mouse/cursor.dart';
@@ -82,11 +81,11 @@ export 'engine/mouse/prevent_default.dart';
 export 'engine/navigation/history.dart';
 export 'engine/noto_font.dart';
 export 'engine/noto_font_encoding.dart';
+export 'engine/occlusion_map.dart';
 export 'engine/onscreen_logging.dart';
 export 'engine/platform_dispatcher.dart';
 export 'engine/platform_dispatcher/app_lifecycle_state.dart';
 export 'engine/platform_dispatcher/view_focus_binding.dart';
-export 'engine/platform_views.dart';
 export 'engine/platform_views/content_manager.dart';
 export 'engine/platform_views/embedder.dart';
 export 'engine/platform_views/message_handler.dart';
@@ -99,9 +98,6 @@ export 'engine/profiler.dart';
 export 'engine/raw_keyboard.dart';
 export 'engine/renderer.dart';
 export 'engine/safe_browser_api.dart';
-export 'engine/scene_builder.dart';
-export 'engine/scene_painting.dart';
-export 'engine/scene_view.dart';
 export 'engine/semantics/accessibility.dart';
 export 'engine/semantics/alert.dart';
 export 'engine/semantics/checkable.dart';
