+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/picture_recorder.dart
@@ -7,7 +7,7 @@ import 'dart:typed_data';
 import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart' as ui;
 
-class CkPictureRecorder implements LayerPictureRecorder, ScenePictureRecorder {
+class CkPictureRecorder implements LayerPictureRecorder {
   SkPictureRecorder? _skRecorder;
   CkCanvas? _recordingCanvas;
