+++ engine/src/flutter/lib/web_ui/lib/src/engine/canvaskit/renderer.dart
@@ -426,71 +426,6 @@ class CanvasKitRenderer extends Renderer {
   ui.ParagraphBuilder createParagraphBuilder(ui.ParagraphStyle style) =>
       isExperimentalWebParagraph ? WebParagraphBuilder(style) : CkParagraphBuilder(style);
 
-  // TODO(harryterkelsen): Merge this logic with the async logic in
-  // [EngineScene], https://github.com/flutter/flutter/issues/142072.
-  @override
-  Future<void> renderScene(ui.Scene scene, EngineFlutterView view) async {
-    assert(
-      rasterizers.containsKey(view.viewId),
-      "Unable to render to a view which hasn't been registered",
-    );
-    final ViewRasterizer rasterizer = rasterizers[view.viewId]!;
-    final RenderQueue renderQueue = rasterizer.queue;
-    final FrameTimingRecorder? recorder = FrameTimingRecorder.frameTimingsEnabled
-        ? FrameTimingRecorder()
-        : null;
-    if (renderQueue.current != null) {
-      // If a scene is already queued up, drop it and queue this one up instead
-      // so that the scene view always displays the most recently requested scene.
-      renderQueue.next?.completer.complete();
-      final Completer<void> completer = Completer<void>();
-      renderQueue.next = (scene: scene, completer: completer, recorder: recorder);
-      return completer.future;
-    }
-    final Completer<void> completer = Completer<void>();
-    renderQueue.current = (scene: scene, completer: completer, recorder: recorder);
-    unawaited(_kickRenderLoop(rasterizer));
-    return completer.future;
-  }
-
-  Future<void> _kickRenderLoop(ViewRasterizer rasterizer) async {
-    final RenderQueue renderQueue = rasterizer.queue;
-    final RenderRequest current = renderQueue.current!;
-    try {
-      await _renderScene(current.scene, rasterizer, current.recorder);
-      current.completer.complete();
-    } catch (error, stackTrace) {
-      current.completer.completeError(error, stackTrace);
-    }
-    renderQueue.current = renderQueue.next;
-    renderQueue.next = null;
-    if (renderQueue.current == null) {
-      return;
-    } else {
-      return _kickRenderLoop(rasterizer);
-    }
-  }
-
-  Future<void> _renderScene(
-    ui.Scene scene,
-    ViewRasterizer rasterizer,
-    FrameTimingRecorder? recorder,
-  ) async {
-    // "Build finish" and "raster start" happen back-to-back because we
-    // render on the same thread, so there's no overhead from hopping to
-    // another thread.
-    //
-    // CanvasKit works differently from the HTML renderer in that in HTML
-    // we update the DOM in SceneBuilder.build, which is these function calls
-    // here are CanvasKit-only.
-    recorder?.recordBuildFinish();
-    recorder?.recordRasterStart();
-
-    await rasterizer.draw((scene as LayerScene).layerTree, null);
-    recorder?.recordRasterFinish();
-    recorder?.submitTimings();
-  }
-
   @override
   void clearFragmentProgramCache() {
     _programs.clear();
@@ -533,5 +468,14 @@ class CanvasKitRenderer extends Renderer {
   );
 
   @override
-  void dumpDebugInfo() {}
+  void dumpDebugInfo() {
+    int i = 0;
+    for (final viewRasterizer in rasterizers.values) {
+      final Map<String, dynamic>? debugJson = viewRasterizer.dumpDebugInfo();
+      if (debugJson != null) {
+        downloadDebugInfo('flutter-scene$i', debugJson);
+        i++;
+      }
+    }
+  }
 }
