+++ engine/src/flutter/lib/web_ui/lib/src/engine/skwasm/skwasm_impl.dart
@@ -14,6 +14,7 @@ export 'skwasm_impl/filters.dart';
 export 'skwasm_impl/font_collection.dart';
 export 'skwasm_impl/image.dart';
 export 'skwasm_impl/memory.dart';
+export 'skwasm_impl/offscreen_canvas_rasterizer.dart';
 export 'skwasm_impl/paint.dart';
 export 'skwasm_impl/paragraph.dart';
 export 'skwasm_impl/path.dart';
