+++ engine/src/flutter/lib/web_ui/lib/src/engine/skwasm/skwasm_impl/picture.dart
@@ -6,7 +6,7 @@ import 'package:ui/src/engine.dart';
 import 'package:ui/src/engine/skwasm/skwasm_impl.dart';
 import 'package:ui/ui.dart' as ui;
 
-class SkwasmPicture extends SkwasmObjectWrapper<RawPicture> implements ScenePicture {
+class SkwasmPicture extends SkwasmObjectWrapper<RawPicture> implements LayerPicture {
   SkwasmPicture.fromHandle(PictureHandle handle) : super(handle, _registry);
 
   static final SkwasmFinalizationRegistry<RawPicture> _registry =
@@ -39,7 +39,7 @@ class SkwasmPicture extends SkwasmObjectWrapper<RawPicture> implements ScenePict
 }
 
 class SkwasmPictureRecorder extends SkwasmObjectWrapper<RawPictureRecorder>
-    implements ScenePictureRecorder {
+    implements LayerPictureRecorder {
   SkwasmPictureRecorder() : super(pictureRecorderCreate(), _registry);
 
   static final SkwasmFinalizationRegistry<RawPictureRecorder> _registry =
