+++ engine/src/flutter/lib/web_ui/lib/src/engine/skwasm/skwasm_impl/raw/raw_surface.dart
@@ -33,6 +33,12 @@ external void surfaceSetCallbackHandler(SurfaceHandle surface, OnRenderCallbackH
 @Native<Void Function(SurfaceHandle)>(symbol: 'surface_destroy', isLeaf: true)
 external void surfaceDestroy(SurfaceHandle surface);
 
+@Native<Void Function(SurfaceHandle, Int)>(
+  symbol: 'surface_setResourceCacheLimitBytes',
+  isLeaf: true,
+)
+external void surfaceSetResourceCacheLimitBytes(SurfaceHandle surface, int bytes);
+
 @Native<Int32 Function(SurfaceHandle, Pointer<PictureHandle>, Int, Int, Int)>(
   symbol: 'surface_renderPictures',
   isLeaf: true,
