+++ engine/src/flutter/lib/web_ui/skwasm/surface.h
@@ -58,6 +58,7 @@ class Surface {
 
   // Main thread only
   void dispose();
+  void setResourceCacheLimit(int bytes);
   uint32_t renderPictures(flutter::DisplayList** picture,
                           int width,
                           int height,
